refactor(OOP): 서비스 레이어에서 리포지토리 구현체 대신 인터페이스를 사용하게 하여 구현체 간의 강한 결합 분리
