+++ src/main/java/onboarding/Problem6.java
@@ -1,13 +1,14 @@
 package onboarding;
 
 import onboarding.problem6.repository.CrewRepository;
+import onboarding.problem6.repository.CrewRepositoryImpl;
 import onboarding.problem6.service.CrewService;
 
 import java.util.List;
 
 public class Problem6 {
     public static List<String> solution(List<List<String>> forms) {
-        CrewRepository crewRepository = new CrewRepository();
+        CrewRepository crewRepository = new CrewRepositoryImpl();
         CrewService crewService = new CrewService(crewRepository);
 
         crewService.saveAllCrews(forms);
