+++ src/main/java/onboarding/problem6/service/CrewService.java
@@ -4,7 +4,9 @@ import onboarding.problem6.repository.CrewRepository;
 import onboarding.problem6.validation.CrewValidator;
 import onboarding.problem6.vo.Crew;
 
-import java.util.*;
+import java.util.List;
+import java.util.Set;
+import java.util.TreeSet;
 import java.util.stream.Collectors;
 
 /**
